+++ src/cmprof.h
@@ -6,14 +6,14 @@
 class CMProf
 	{
 public:
-	vector<vector<double> > m_MeanDistMx;
-	vector<vector<double> > m_StdDevs;
+	vector<vector<float> > m_MeanDistMx;
+	vector<vector<float> > m_StdDevs;
 	const SeqDB *m_MSA = 0;
 
 // Training only
 	vector<bool> m_ColIsCore;
 	vector<uint> m_CoreCols;
-	vector<vector<vector<double> > > m_DistMxVec;
+	vector<vector<vector<float> > > m_DistMxVec;
 	map<string, uint> m_UngappedSeqToIdx;
 
 
@@ -37,9 +37,9 @@ public:
 
 	void ToFile(const string &FileName) const;
 	void MxToFile(FILE *f, const string &Name,
-	  const vector<vector<double> > &Mx) const;
+	  const vector<vector<float> > &Mx) const;
 	void MxFromFile(FILE *f, string &Name, uint CoreColCount,
-	  vector<vector<double> > &Mx);
+	  vector<vector<float> > &Mx);
 	void FromFile(FILE *f);
 	void FromFile(const string &FileName);
 
@@ -53,9 +53,9 @@ public:
 	void FinalizeTrain();
 	bool TrainChain(const PDBChain &Chain);
 	void GetDistMx(const PDBChain &Chain, const vector<uint> &PosVec,
-	  vector<vector<double> > &DistMx);
+	  vector<vector<float> > &DistMx);
 	void GetMeanStdDev(uint i, uint j,
-	  double &Mean, double &StdDev) const;
+	  float &Mean, float &StdDev) const;
 	};
 
-double GetNormal(double Mu, double Sigma, double x);
+float GetNormal(float Mu, float Sigma, float x);
