+++ src/valuetoint.cpp
@@ -3,7 +3,7 @@
 
 #define BIN_T(Feat, Idx, t)	if (Value < t) return Idx;
 
-uint DSS::ValueToInt_NENDist(double Value) const
+uint DSS::ValueToInt_NENDist(float Value) const
 	{
 	if (Value < 4.417) return 0;
 	if (Value < 4.647) return 1;
@@ -23,7 +23,7 @@ uint DSS::ValueToInt_NENDist(double Value) const
 	return 15;
 	}
 
-uint DSS::ValueToInt_RENDist(double Value) const
+uint DSS::ValueToInt_RENDist(float Value) const
 	{
 	if (Value < 6) return 0;
 	if (Value < 7) return 1;
@@ -43,7 +43,7 @@ uint DSS::ValueToInt_RENDist(double Value) const
 	return 15;
 	}
 
-uint DSS::ValueToInt_DstNxtHlx(double Value) const
+uint DSS::ValueToInt_DstNxtHlx(float Value) const
 	{
 	if (Value < 6) return 0;
 	if (Value < 7) return 1;
@@ -63,7 +63,7 @@ uint DSS::ValueToInt_DstNxtHlx(double Value) const
 	return 15;
 	}
 
-uint DSS::ValueToInt_StrandDens(double Value) const
+uint DSS::ValueToInt_StrandDens(float Value) const
 	{
 	if (Value < 0.02212) return 0;
 	if (Value < 0.07567) return 1;
@@ -83,7 +83,7 @@ uint DSS::ValueToInt_StrandDens(double Value) const
 	return 15;
 	}
 
-uint DSS::ValueToInt_NormDens(double Value) const
+uint DSS::ValueToInt_NormDens(float Value) const
 	{
 	if (Value < 0.241) return 0;
 	if (Value < 0.3399) return 1;
@@ -103,7 +103,7 @@ uint DSS::ValueToInt_NormDens(double Value) const
 	return 15;
 	}
 
-uint DSS::ValueToInt_HelixDens(double Value) const
+uint DSS::ValueToInt_HelixDens(float Value) const
 	{
 BIN_T(HelixDens, 0, 0.03015);
 BIN_T(HelixDens, 1, 0.06112);
@@ -123,7 +123,7 @@ BIN_T(HelixDens, 14, 0.6343);
 	return 15;
 	}
 
-uint DSS::ValueToInt_PMDist(double Value) const
+uint DSS::ValueToInt_PMDist(float Value) const
 	{
 BIN_T(PMDist, 0, 9.994);
 BIN_T(PMDist, 1, 12.06);
@@ -143,7 +143,7 @@ BIN_T(PMDist, 14, 32.77);
 	return 15;
 	}
 
-uint DSS::ValueToInt_DstPrvHlx(double Value) const
+uint DSS::ValueToInt_DstPrvHlx(float Value) const
 	{
 BIN_T(DstNxtHlx, 0, 0);
 BIN_T(DstNxtHlx, 1, 6);
@@ -163,7 +163,7 @@ BIN_T(DstNxtHlx, 14, 28.82);
 	return 15;
 	}
 
-uint DSS::ValueToInt_NX(double Value) const
+uint DSS::ValueToInt_NX(float Value) const
 	{
 BIN_T(NX, 0, 20.65);
 BIN_T(NX, 1, 23.54);
